Fixed problems in for loops

+++ install.lua
@@ -1,7 +1,7 @@
 --[[
   Author: Panzer1119
   
-  Date: Edited 25 Jun 2018 - 08:34 PM
+  Date: Edited 25 Jun 2018 - 08:38 PM
   
   Original Source: https://github.com/Panzer1119/CCUtils/blob/master/install.lua
   
@@ -19,25 +19,29 @@ f.close()
 
 installed = true
 
-for (i = 1, #files) do
+for i = 1, #files do
 	print("")
 	local download = true
 	local file = files[i]
-	if (fs.exists(file.path)) then
-		download = false
-		print(file.name .. " is already installed,")
-		write("do you want to update it? (Y/n): ")
-		local input = read()
-		if (input == "Y") then
-			fs.delete(file.path)
-			download = true
+	if (file == nil or file.name == nil or file.path == nil or file.url == nil) then
+		print("Error with a file!")
+	else
+		if (fs.exists(file.path)) then
+			download = false
+			print(file.name .. " is already installed,")
+			write("do you want to update it? (Y/n): ")
+			local input = read()
+			if (input == "Y") then
+				fs.delete(file.path)
+				download = true
+			end
 		end
-	end
-	if (download) then
-		print("Downloading " .. file.name .. " to " .. file.path .. "...")
-		shell.run("wget " .. file.url .. " " .. file.path)
-		if not (fs.exists(file.path)) then
-			installed = false
+		if (download) then
+			print("Downloading " .. file.name .. " to " .. file.path .. "...")
+			shell.run("wget " .. file.url .. " " .. file.path)
+			if not (fs.exists(file.path)) then
+				installed = false
+			end
 		end
 	end
 end
@@ -57,9 +61,9 @@ else
 	print("")
 	print("")
 	print("Some files failed and could not be installed,")
-	write("do you want to try it again? (Y/n): ")
+	write("do you want to try it again? (yes/n): ")
 	local input = read()
-	if (input == "Y") then
+	if (input == "yes") then
 		shell.run(shell.getRunningProgram())
 	end
 end
